nit: move `WithOtherFields` out of eth namespace
