+++ crates/rpc-types/src/lib.rs
@@ -23,3 +23,6 @@ mod eth;
 
 pub use alloy_serde as serde_helpers;
 pub use eth::*;
+
+mod with_other;
+pub use with_other::WithOtherFields;
